Major refactoring of modules
Instead of having the model handle both power supplies, instead we're doing one model, one supply, which simplifies down a lot of code on the model side and doesn't complicate things much in the controller.

Display continues to change, and I'm including the `testDisplay.py` file I'm using so that I don't have to worry about TCP connections while messing with it. Note that this file is temporary and PEP8 compliant, just to see how it looks.